% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mock.R
\name{mock}
\alias{mock}
\title{Stubbing}
\usage{
mock(f, replacement)
}
\arguments{
\item{f}{a function whose introspection is watched.}

\item{replacement}{function (or value) list to replace.}
}
\description{
Replaces function calls in functions.
}
\examples{
f <- function(x, y) {
   a <- g(x)
   b <- h(x, y)
   a + b
}
sf <- mock(f, list(g=1, h=2))
sf()
}

