% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/substitute.R
\name{substitute.variable}
\alias{substitute.variable}
\title{\code{substitute} equivalent, but an expression parameter is passed as a variable}
\usage{
substitute.variable(var, env)
}
\arguments{
\item{var}{an expression variable}

\item{env}{substitution rule}
}
\description{
\code{substitute} equivalent, but an expression parameter is passed as a variable
}

