% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mock.R
\name{mock}
\alias{mock}
\title{Mocking}
\usage{
mock(f, verify, violate = stopifnot)
}
\arguments{
\item{f}{a function whose introspection is watched.}

\item{verify}{constraint that each variable or function call to be satisfied given as a list}

\item{violate}{action when the verification is not passed.}
}
\description{
Watches variables and function calls to check they are used correctly.
}
\examples{
f <- function(x, y) x + y
mf <- mock(f, list(x = quote(x > 2)))
mf(3, 1)
\dontrun{
mf(1, 2)
}

g <- function(z) f(z + 1, z - 1)
mg <- mock(g, list(f = list(x = quote(x > 1))))
mg(1)
\dontrun{
mg(0)
}
}

